Move getInterpreterInfo() to the right file.
