+++ src/client/pythonEnvironments/base/locators/composite/resolvingLocator.ts
@@ -6,7 +6,7 @@ import { Event, EventEmitter } from 'vscode';
 import { traceVerbose } from '../../../../common/logger';
 import { IEnvironmentInfoService } from '../../../info/environmentInfoService';
 import { PythonEnvInfo } from '../../info';
-import { InterpreterInformation } from '../../info/interpreter';
+import { InterpreterInformation } from '../../../info/tool';
 import {
     ILocator, IPythonEnvsIterator, PythonEnvUpdatedEvent, PythonLocatorQuery,
 } from '../../locator';
